requirements.txt
chainlit==1.0.0
google-cloud-bigquery==3.11.4
